chore: rename jax.py to avoid circular imports
